fixed 3a6ff25 from https://gitee.com/zzb999_admin/arkcompiler_ets_runtime_3.2beta3_2/pulls/2368 fix codex and codeCheck warnings

+++ ecmascript/object_factory.cpp
@@ -219,7 +219,7 @@ void ObjectFactory::NewJSArrayBufferData(const JSHandle<JSArrayBuffer> &array, i
     }
 
     JSTaggedValue data = array->GetArrayBufferData();
-    size_t size = length * sizeof(uint8_t);
+    size_t size = static_cast<size_t>(length) * sizeof(uint8_t);
     if (data != JSTaggedValue::Undefined()) {
         auto *pointer = JSNativePointer::Cast(data.GetTaggedObject());
         auto newData = vm_->GetNativeAreaAllocator()->AllocateBuffer(size);
@@ -247,7 +247,8 @@ void ObjectFactory::NewJSSharedArrayBufferData(const JSHandle<JSArrayBuffer> &ar
         return;
     }
     void *newData = nullptr;
-    size_t size = JSSharedMemoryManager::GetInstance()->CreateOrLoad(&newData, length) ? length : 0;
+    size_t size =
+        JSSharedMemoryManager::GetInstance()->CreateOrLoad(&newData, length) ? static_cast<size_t>(length) : 0U;
     if (memset_s(newData, length, 0, length) != EOK) {
         LOG_FULL(FATAL) << "memset_s failed";
         UNREACHABLE();
